requirements.txt
cement==3.0.4
pyyaml~=5.3.1
networkx~=2.5
tabulate~=0.8.7

setuptools>=38.6.0
urllib3~=1.25.8
confuse~=1.3.0
certifi >= 14.05.14
six >= 1.10
python_dateutil >= 2.5.3

pytest~=6.1.1
pylint~=2.6.0
pytest-pylint~=0.18.0
bump2version~=1.0.1
gitpython~=3.1.11
docker~=4.1.0
